Fix: The children of non-clipping viewees are clipped when drawn.
This was caused by a chance to the rendering mechanism whereby the
damaged region is clipped before rendering.

But when calculating the damaged regions we did not account for
non-clipping viewees, and the damaged region (and thus clip area) was
set to the to non-clipping viewee - resulting in its children being
clipped.

To combat this, we now add to the damaged area the bounds of the
children of a non-clipping viewee. We do so recursively (so if the
child isn’t clipping its children bounds will be added to the damaged
area as well).