Fix: Rendering children outside the bound of a non-clipping viewee aren't rendered.
Previously, we were checking that the children of non-clipping viewees
are rendered, but not the case where the child viewee is completely
outside the bounds of its parent.

Because isWithinClipArea() was only checking against the bounds of the
viewee, the following happened:

* A child viewee would update (with its bounds completely outside the
parent bounds).
* When rendering, isWithinClipArea() would return false when checking
the parent, and so the child won’t render.

To fix this, isWithinClipArea() now checks agains the bounds of the
non-clipping composition starting at the viewee provided.