Fix: Antialiased strokes are not redrawn after erase.
When a figure is updated, its bounds are damaged and become the clip
area for the next rendering event. But if only the antialiased stroke
of another viewee is within such clip area (and not the viewee bounds
itself), the antialiased area does not refresh.

To fix this, we now account for antialiasing margin when checking if a
viewee is within the clip area.